refactor: removes boolector_impl from sat_solver.rs and fix imports
